+++ resources/lang/pt/grid.php
@@ -1,0 +1,11 @@
+<?php
+return [
+    'page-info' => 'Exibindo :start-:end de :total itens',
+    'select' => 'Selecione',
+    'actions' => 'Ações',
+    'delete' => 'Excluir',
+    'deletion' => 'Exclusão',
+    'search' => 'Buscar',
+    'reset' => 'Limpar',
+    'send' => 'Enviar',
+];