+++ changelog.md
@@ -1,5 +1,10 @@
 ### CHANGE LOG:
 
+**1.0.12 April 12, 2023:**
+- Add custom class and style to filter buttons
+- Options to show count column
+- Add pt translation
+
 **1.0.11 February 18, 2023:**
 - Add support for Laravel 10